Windows install updated and some fixes (#4)
The windows install script has been made more resistant to possible
  interfering existing tool installations and faulty path variable.
New uart name parameter has been added to the do command. This can be used
  if the automatic detection of the XPLR-IOT-1 serial port fails.
The extFsFree function now return possible error code when applicable.
The socket example now has the correct echo server address.

+++ examples/common/ext_fs.h
@@ -37,9 +37,9 @@ const char *extFsPath(const char *fileName);
 
 /**
  * Get the size of the free space on the file system in kB.
- * @return Actual free size.
+ * @return Actual free size or possible negative error code
  */
-unsigned long extFsFree();
+int32_t extFsFree();
 
 /**
  * Check if a file exists
